Adds the GBS device, program_utils module.
