+++ tests/frontend/test_program.py
@@ -26,7 +26,6 @@ from strawberryfields import ops
 
 from strawberryfields import devicespecs
 from strawberryfields.devicespecs.device_specs import DeviceSpecs
-from strawberryfields.temporary import check_GBS
 
 
 # make test deterministic
@@ -611,8 +610,8 @@ class TestValidation:
                 new_prog = prog.compile(backend='dummy')
 
 
-class TestGBSCheck:
-    """Test the Gaussian boson sampling check method."""
+class TestGBS:
+    """Test the Gaussian boson sampling device."""
 
     def test_GBS_fail_1(self):
         """GBS failure."""
@@ -622,20 +621,9 @@ class TestGBSCheck:
             ops.Rgate(1.0)  | q[0]
 
         with pytest.raises(program.CircuitError, match="Operations following the Fock measurements."):
-            check_GBS(prog)
+            prog.compile('gbs')
 
     def test_GBS_fail_2(self):
-        """GBS failure."""
-        prog = sf.Program(2)
-        with prog.context as q:
-            ops.Sgate(-0.3) | q[1]
-            ops.Kgate(1.0) | q[0]
-            ops.Measure | q
-
-        with pytest.raises(program.CircuitError, match="Non-gaussian Operation:"):
-            check_GBS(prog)
-
-    def test_GBS_fail_3(self):
         """GBS failure."""
         prog = sf.Program(2)
         with prog.context as q:
@@ -643,9 +631,9 @@ class TestGBSCheck:
             ops.Sgate(-0.5) | q[1]
 
         with pytest.raises(program.CircuitError, match="No Fock measurements."):
-            check_GBS(prog)
+            prog.compile('gbs')
 
-    def test_GBS_fail_4(self):
+    def test_GBS_fail_3(self):
         """GBS failure."""
         prog = sf.Program(2)
         with prog.context as q:
@@ -656,9 +644,9 @@ class TestGBSCheck:
             ops.Measure | q[1]
 
         with pytest.raises(program.CircuitError, match="The Fock measurements are not consecutive."):
-            check_GBS(prog)
+            prog.compile('gbs')
 
-    def test_GBS_fail_5(self):
+    def test_GBS_fail_4(self):
         """GBS failure."""
         prog = sf.Program(3)
         with prog.context as q:
@@ -667,7 +655,7 @@ class TestGBSCheck:
             ops.Measure | q
 
         with pytest.raises(program.CircuitError, match="Measuring the same mode more than once."):
-            check_GBS(prog)
+            prog.compile('gbs')
 
     def test_GBS_success(self):
         """GBS check passes."""
@@ -680,7 +668,7 @@ class TestGBSCheck:
             ops.Rgate(-1.0) | q[1]
             ops.Measure | q[1]
 
-        check_GBS(prog)
+        prog = prog.compile('gbs')
         assert len(prog.circuit) == 4
         assert prog.circuit[-1].op.__class__ == ops.MeasureFock
         assert [x.ind for x in prog.circuit[-1].reg] == list(range(3))
